Add tool for programatically downloading MyTurn data
The tool is wrapped in a CLI using Click, a handy Python CLI library.
It is designed in a way that isn't too "CHTL" specific - you can
override the tool library being downloaded.

This also moves the anonymization script to the CLI directory and makes
it less opinionated about where data needs to live.