Add a bot for accessign MyTurn, and a script for clearing PII
The MyTurnBot is nice because it encapsulates startingt a session,
saving cookies,  sending to the URL, etc.  I think this will be one
of the more useful parts of this sytem, and should be extended to
support more parts of the MyTurn API which don't have a real API.

The personal info clearer is meant to remove PII collected previously.
To use it, make sure to unmark any fields that should be cleared as
"Required" in "Settings > Membership > Registration".